requirements.txt
google-cloud-storage==1.38.0
google-cloud-secret-manager==2.5.0
TwitterAPI==2.7.4
dacite==1.6.0
